+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## v0.8.1
+
+### New function
+
+Added a new function `with_multisample()` in `BrushBuilder` by @AndriBaal [#12](https://github.com/Blatko1/wgpu-text/pull/12) which specifies the `wgpu::MultisampleState` for the inner pipeline.
+
+### Minor changes
+
+- changed some inner (private) variable names
+
 ## v0.8.0
 
 ### Breaking changes